Clock client load time reduced
